Hid snippet management UI from staff users without permissions
no issue

- snippets can only be created and deleted by owners/admins/editors
- added a property in the editor controller to determine if the logged in user has sufficient permissions, then only pass the appropriate save/delete snippet actions to the editor component if the check is passed
- updates koenig menus and toolbars to skip rendering of buttons if the associated action function is not available

+++ ghost/admin/lib/koenig-editor/addon/components/koenig-slash-menu.js
@@ -68,17 +68,20 @@ export default class KoenigSlashMenuComponent extends Component {
             };
 
             snippets.forEach((snippet) => {
-                snippetsSection.items.push({
+                let snippetItem = {
                     label: snippet.name,
                     icon: snippetIcon(snippet),
                     type: 'snippet',
-                    matches: [snippet.name.toLowerCase()],
-                    deleteClicked: (event) => {
+                    matches: [snippet.name.toLowerCase()]
+                };
+                if (this.args.deleteSnippet) {
+                    snippetItem.deleteClicked = (event) => {
                         event.preventDefault();
                         event.stopImmediatePropagation();
                         this.args.deleteSnippet(snippet);
-                    }
-                });
+                    };
+                }
+                snippetsSection.items.push(snippetItem);
             });
 
             itemSections.push(snippetsSection);
